add user auth for offer operations

+++ app/Http/Controllers/OfferController.php
@@ -40,7 +40,13 @@ class OfferController extends Controller
     }
     public function edit($id){
         $offer = Offer::find($id);
-        return view('offer.edit')->with('offer', $offer);
+        if($offer->user_id == auth()->user()->id){
+            return view('offer.edit')->with('offer', $offer);
+        }
+        else{
+            return redirect('https://www.youtube.com/watch?v=73T5NVNb7lE');
+        }
+        
     }
 
     public function update(Request $request, $id){
@@ -48,14 +54,23 @@ class OfferController extends Controller
             'name' => ['required', 'max:255'],
             'description' => ['required']
         ]);
-
         $offer = Offer::find($id);
+        if($offer->user_id == auth()->user()->id){
         $offer->update($request->all());
         return $offer;
+        }
+        else{
+            return redirect('https://www.youtube.com/watch?v=73T5NVNb7lE');
+        }
     }
     public function destroy($id){
         $offer = Offer::find($id);
+        if($offer->user_id == auth()->user()->id){
         $offer->delete();
         return 'succes';
+        }
+        else{
+            return redirect('https://www.youtube.com/watch?v=73T5NVNb7lE');
+        }
     }
 }
